refactor local game
