+++ frontend/src/pong/game/local_game.js
@@ -3,77 +3,16 @@ import { showMessage } from "../../utils/utils.js"
 import { handleRoute, navigate } from "../../routes/route_helpers.js";
 import { currentEngine, disposeCurrentEngine, resizeListener } from "../../routes/routes.js";
 import { MeshBuilder, StandardMaterial, Color3, Vector3 } from "@babylonjs/core";
+import defaultPhysics from "./physicsConfig.js";
+import { createPaddleVisual } from "./paddleVisuals.js";
+import { createTableArena, positionPaddles } from "./arenaBuilder.js";
+import { createPhysicsPanel } from "./physicsPanel.js";
 
 export function initOfflineGame(scene, gameObjects, tournament) {
     return new Promise((resolve) => {
         // These values are the live tuning knobs for the offline 3D pong model.
         // I keep them grouped so I can reason about ball travel, bounce feel, spin,
         // and paddle motion without hunting through the frame loop.
-        const defaultPhysics = {
-            initialSpeedX: 0.10,
-            initialLateralSpeedZ: 0.013,
-            initialUpwardSpeedY: 0.09,
-            speedMultiplier: 0.999,
-            speedMax: 0.24,
-            gravity: 0.0042,
-            magnusCoefficient: 0.0008,
-            magnusExtraFactor: 20,
-            spinDecay: 0.992,
-            visualSpinFactor: 100,
-            tableTopY: 0,
-            tableHalfLength: 5.9,
-            tableHalfWidth: 2.95,
-            // Table is slightly higher at center and lower toward each end.
-            tableInclinePerUnitX: 0.01,
-            tableDownhillAcceleration: 0.0016,
-            tableBounceRestitution: 0.9,
-            tableBounceFriction: 0.985,
-            sideWallRestitution: 0.92,
-            sideWallSpinRetention: 0.9,
-            // Net height controls the center barrier visually and physically.
-            netHeight: 0.2,
-            netHalfThickness: 0.05,
-            netBounceDamping: 0.6,
-            netSpinRetention: 0.75,
-            floorY: -2,
-            goalX: 6.3,
-            // Collision dimensions are tuned to match the visible paddle face,
-            // plus a small margin for the ball radius.
-            paddleCollisionDepth: 0.2,
-            paddleCollisionHeight: 0.9,
-            paddleCollisionWidth: 0.45,
-            // Sideways paddle motion feeds these spin/trajectory values on contact.
-            spinTransferCoefficient: 1.5,
-            velocityTransfer: 0.11,
-            paddleBounceBoost: 1.02,
-            paddleBaseY: 0.9,
-            paddleBoundsZ: 2.5,
-            paddleRotationStep: 0.06,
-            paddleRotationLimit: 0.65,
-            // Tilt and rotation speed can add vertical lift to the outgoing ball.
-            paddleTiltToVerticalVelocity: 0.12,
-            paddleApproachBoost: 0.14,
-            paddleRetreatDampingStrength: 0.2,
-            paddleRetreatDampingMin: 0.65,
-            paddleRotationLiftFromSpeed: 0.02,
-            paddleStepZ: 0.10,
-            paddleSwingReach: 1.62,
-            paddleSwingRotationBoost: 0.86,
-            paddleSwingDuration: 0.24,
-            paddleLobDuration: 0.2,
-            paddleLobLiftBoost: 0.12,
-            paddleLobSlowFactor: 0.5,
-            paddleLobSideDamping: 0.65,
-            paddleLobRotationBoost: 0.22,
-            paddleWingDuration: 0.22,
-            paddleWingReach: 0.38,
-            paddleWingSideReach: 0.42,
-            paddleWingLiftBoost: 0.16,
-            paddleWingSideBoost: 0.08,
-            paddleWingSpinBoost: 0.85,
-            paddleWingForwardFactor: 0.95,
-            paddleWingRotationBoost: 0.28,
-        };
         const physics = { ...defaultPhysics };
 
         // The paddles are placed at the two ends of the table on the X axis.
@@ -124,36 +63,9 @@ export function initOfflineGame(scene, gameObjects, tournament) {
         const clamp = (value, min, max) => Math.max(min, Math.min(max, value));
         const randomSign = () => (Math.random() > 0.5 ? 1 : -1);
 
-        const createPaddleVisual = (paddle, faceColor) => {
-            // I build the visual paddle as a separate mesh so the collision body can
-            // stay simple while the rendered shape looks closer to a real paddle.
-            const faceMat = new StandardMaterial(`${paddle.name}FaceMat`, scene);
-            faceMat.diffuseColor = faceColor;
+        // paddle visuals handled in ./paddleVisuals.js
 
-            // A real paddle is mostly a flat square face, so I model it as a thin slab.
-            const paddleFace = MeshBuilder.CreateBox(`${paddle.name}Face`, {
-                width: 0.55,
-                height: 1.55,
-                depth: 0.12,
-            }, scene);
-            // Rotate the face 90 degrees so the thin side points toward the ball.
-            paddleFace.rotation.y = Math.PI / 2;
-            paddleFace.parent = paddle;
-
-            paddleFace.material = faceMat;
-            paddle.isVisible = false;
-            return { paddleFace };
-        };
-
-        const positionPaddles = () => {
-            // Each paddle starts locked to its end of the table, at the same height.
-            gameObjects.paddleLeft.position.x = paddleLeftBaseX;
-            gameObjects.paddleRight.position.x = paddleRightBaseX;
-            gameObjects.paddleLeft.position.y = physics.paddleBaseY + 2;
-            gameObjects.paddleRight.position.y = physics.paddleBaseY;
-            gameObjects.paddleLeft.position.z = clamp(gameObjects.paddleLeft.position.z, -physics.paddleBoundsZ, physics.paddleBoundsZ);
-            gameObjects.paddleRight.position.z = clamp(gameObjects.paddleRight.position.z, -physics.paddleBoundsZ, physics.paddleBoundsZ);
-        };
+        // paddle positioning handled in ./arenaBuilder.js
 
         const resetBall = (serveDirection = 1) => {
             // I restart the rally from the middle of the table so the serve is predictable.
@@ -166,71 +78,9 @@ export function initOfflineGame(scene, gameObjects, tournament) {
             ballVZ = physics.initialLateralSpeedZ * randomSign();
         };
 
-        const createTableArena = () => {
-            // Table surface, white boundary lines, a center net, and a dark floor plane
-            // give the scene a proper table-tennis feel instead of floating objects.
-            const tableMat = new StandardMaterial("localTableMat", scene);
-            tableMat.diffuseColor = new Color3(0.06, 0.22, 0.36);
+        // arena creation handled in ./arenaBuilder.js
 
-            const lineMat = new StandardMaterial("localTableLineMat", scene);
-            lineMat.diffuseColor = new Color3(0.95, 0.95, 0.95);
-
-            const netMat = new StandardMaterial("localNetMat", scene);
-            netMat.diffuseColor = new Color3(0.92, 0.92, 0.95);
-
-            const floorMat = new StandardMaterial("localFloorMat", scene);
-            floorMat.diffuseColor = new Color3(0.08, 0.1, 0.14);
-
-            const table = MeshBuilder.CreateBox("localPongTable", {
-                width: physics.tableHalfLength * 2,
-                height: 0.16,
-                depth: physics.tableHalfWidth * 2,
-            }, scene);
-            table.position = new Vector3(0, physics.tableTopY - 0.08, 0);
-            table.material = tableMat;
-
-            const centerLine = MeshBuilder.CreateBox("localPongCenterLine", {
-                width: physics.tableHalfLength * 2,
-                height: 0.01,
-                depth: 0.02,
-            }, scene);
-            centerLine.position = new Vector3(0, physics.tableTopY + 0.085, 0);
-            centerLine.material = lineMat;
-
-            const sideLineLeft = MeshBuilder.CreateBox("localPongSideLineLeft", {
-                width: physics.tableHalfLength * 2,
-                height: 0.01,
-                depth: 0.02,
-            }, scene);
-            sideLineLeft.position = new Vector3(0, physics.tableTopY + 0.085, physics.tableHalfWidth - 0.06);
-            sideLineLeft.material = lineMat;
-
-            const sideLineRight = MeshBuilder.CreateBox("localPongSideLineRight", {
-                width: physics.tableHalfLength * 2,
-                height: 0.01,
-                depth: 0.02,
-            }, scene);
-            sideLineRight.position = new Vector3(0, physics.tableTopY + 0.085, -physics.tableHalfWidth + 0.06);
-            sideLineRight.material = lineMat;
-
-            // Build net with unit height so slider changes can resize it safely at runtime.
-            const net = MeshBuilder.CreateBox("localPongNet", {
-                width: physics.netHalfThickness * 2,
-                height: 1,
-                depth: physics.tableHalfWidth * 2 - 0.08,
-            }, scene);
-            net.scaling.y = physics.netHeight;
-            net.position = new Vector3(0, physics.tableTopY + physics.netHeight / 2, 0);
-            net.material = netMat;
-
-            const floor = MeshBuilder.CreateGround("localPongFloor", { width: 22, height: 14 }, scene);
-            floor.position.y = physics.floorY;
-            floor.material = floorMat;
-
-            return { table, centerLine, sideLineLeft, sideLineRight, net, floor };
-        };
-
-        positionPaddles();
+        positionPaddles(gameObjects, physics, paddleLeftBaseX, paddleRightBaseX);
         resetBall(1);
 
         if (scene.activeCamera) {
@@ -241,9 +91,9 @@ export function initOfflineGame(scene, gameObjects, tournament) {
             scene.activeCamera.setTarget(new Vector3(0, physics.tableTopY + -1.0, 0));
         }
 
-        const arenaMeshes = createTableArena();
-        const leftPaddleVisual = createPaddleVisual(gameObjects.paddleLeft, new Color3(0.9, 0.2, 0.18));
-        const rightPaddleVisual = createPaddleVisual(gameObjects.paddleRight, new Color3(0.18, 0.3, 0.95));
+        const arenaMeshes = createTableArena(scene, physics);
+        const leftPaddleVisual = createPaddleVisual(gameObjects.paddleLeft, new Color3(0.9, 0.2, 0.18), scene);
+        const rightPaddleVisual = createPaddleVisual(gameObjects.paddleRight, new Color3(0.18, 0.3, 0.95), scene);
 
         const scoreP1 = document.getElementById("scoreP1");
         const scoreP2 = document.getElementById("scoreP2");
@@ -393,185 +243,9 @@ export function initOfflineGame(scene, gameObjects, tournament) {
             }
         };
 
-        const createPhysicsPanel = () => {
-            // The sliders float over the game so I can tune the match while watching the rally.
-            const gameContainer = document.getElementById("gameContainer") || document.getElementById("app-root");
-            if (!gameContainer) return null;
+        // physics panel provided by ./physicsPanel.js
 
-            const panel = document.createElement("div");
-            panel.id = "localPhysicsPanel";
-            panel.style.position = "absolute";
-            panel.style.left = "16px";
-            panel.style.bottom = "16px";
-            panel.style.zIndex = "35";
-            panel.style.width = "320px";
-            panel.style.maxHeight = "65vh";
-            panel.style.overflowY = "auto";
-            panel.style.padding = "12px";
-            panel.style.border = "1px solid rgba(56, 189, 248, 0.5)";
-            panel.style.borderRadius = "12px";
-            panel.style.background = "rgba(2, 6, 23, 0.8)";
-            panel.style.backdropFilter = "blur(6px)";
-            panel.style.pointerEvents = "auto";
-            panel.style.color = "#e2e8f0";
-            panel.style.fontFamily = "monospace";
-            panel.style.fontSize = "12px";
-            panel.style.userSelect = "none";
-
-            const title = document.createElement("div");
-            title.textContent = "Local Physics";
-            title.style.fontWeight = "700";
-            title.style.marginBottom = "8px";
-            title.style.letterSpacing = "0.04em";
-            panel.appendChild(title);
-
-            const focusHint = document.createElement("div");
-            focusHint.textContent = "Highlighted rows strongly affect side-spin and high arcs";
-            focusHint.style.fontSize = "11px";
-            focusHint.style.color = "#fde68a";
-            focusHint.style.marginBottom = "8px";
-            panel.appendChild(focusHint);
-
-            const dragBar = document.createElement("div");
-            dragBar.textContent = "Drag";
-            dragBar.style.cursor = "move";
-            dragBar.style.marginBottom = "10px";
-            dragBar.style.padding = "6px 8px";
-            dragBar.style.background = "rgba(30, 41, 59, 0.9)";
-            dragBar.style.border = "1px solid rgba(148, 163, 184, 0.45)";
-            dragBar.style.borderRadius = "8px";
-            dragBar.style.textAlign = "center";
-            panel.appendChild(dragBar);
-
-            let isDragging = false;
-            let dragOffsetX = 0;
-            let dragOffsetY = 0;
-
-            const onDragMove = (event) => {
-                if (!isDragging) return;
-                panel.style.left = `${Math.max(0, event.clientX - dragOffsetX)}px`;
-                panel.style.top = `${Math.max(0, event.clientY - dragOffsetY)}px`;
-                panel.style.bottom = "auto";
-            };
-
-            const onDragUp = () => {
-                isDragging = false;
-            };
-
-            dragBar.addEventListener("pointerdown", (event) => {
-                const rect = panel.getBoundingClientRect();
-                isDragging = true;
-                dragOffsetX = event.clientX - rect.left;
-                dragOffsetY = event.clientY - rect.top;
-                dragBar.setPointerCapture(event.pointerId);
-            });
-
-            dragBar.addEventListener("pointermove", onDragMove);
-            dragBar.addEventListener("pointerup", onDragUp);
-            dragBar.addEventListener("pointercancel", onDragUp);
-
-            sliderConfigs.forEach((config) => {
-                const row = document.createElement("div");
-                row.style.marginBottom = "8px";
-                const isHighImpact = highImpactKeys.has(config.key);
-                if (isHighImpact) {
-                    row.style.padding = "6px";
-                    row.style.borderRadius = "8px";
-                    row.style.border = "1px solid rgba(251, 191, 36, 0.6)";
-                    row.style.background = "rgba(120, 53, 15, 0.22)";
-                }
-
-                const topLine = document.createElement("div");
-                topLine.style.display = "flex";
-                topLine.style.justifyContent = "space-between";
-                topLine.style.alignItems = "center";
-
-                const label = document.createElement("label");
-                label.textContent = isHighImpact ? `${config.label} *` : config.label;
-                label.style.color = isHighImpact ? "#fde68a" : "#93c5fd";
-                label.style.fontWeight = isHighImpact ? "700" : "400";
-
-                const value = document.createElement("span");
-                const format = makeValueFormatter(config.step);
-                value.textContent = format(physics[config.key]);
-
-                if (isHighImpact) {
-                    const hint = document.createElement("div");
-                    hint.textContent = highImpactHints[config.key] || "High-impact control";
-                    hint.style.fontSize = "10px";
-                    hint.style.color = "#fde68a";
-                    hint.style.opacity = "0.85";
-                    hint.style.marginBottom = "4px";
-                    row.appendChild(hint);
-                }
-
-                topLine.appendChild(label);
-                topLine.appendChild(value);
-
-                const input = document.createElement("input");
-                input.type = "range";
-                input.min = String(config.min);
-                input.max = String(config.max);
-                input.step = String(config.step);
-                input.value = String(physics[config.key]);
-                input.style.width = "100%";
-
-                input.addEventListener("input", () => {
-                    const parsed = Number(input.value);
-                    physics[config.key] = parsed;
-                    value.textContent = format(parsed);
-
-                    // Keep center net visuals in sync with live Net Height slider value.
-                    if (config.key === "netHeight" && arenaMeshes?.net) {
-                        arenaMeshes.net.scaling.y = parsed;
-                        arenaMeshes.net.position.y = physics.tableTopY + parsed / 2;
-                    }
-                });
-
-                row.appendChild(topLine);
-                row.appendChild(input);
-                panel.appendChild(row);
-            });
-
-            const actions = document.createElement("div");
-            actions.style.display = "flex";
-            actions.style.justifyContent = "flex-end";
-
-            const resetBtn = document.createElement("button");
-            resetBtn.type = "button";
-            resetBtn.textContent = "Reset Defaults";
-            resetBtn.style.padding = "6px 8px";
-            resetBtn.style.borderRadius = "6px";
-            resetBtn.style.border = "1px solid rgba(34, 197, 94, 0.5)";
-            resetBtn.style.background = "rgba(20, 83, 45, 0.8)";
-            resetBtn.style.color = "#dcfce7";
-            resetBtn.style.cursor = "pointer";
-
-            resetBtn.addEventListener("click", () => {
-                sliderConfigs.forEach((config) => {
-                    physics[config.key] = defaultPhysics[config.key];
-                });
-
-                const allInputs = panel.querySelectorAll("input[type='range']");
-                allInputs.forEach((element, index) => {
-                    const cfg = sliderConfigs[index];
-                    const format = makeValueFormatter(cfg.step);
-                    element.value = String(defaultPhysics[cfg.key]);
-                    const valueLabel = element.previousSibling?.lastChild;
-                    if (valueLabel && valueLabel.nodeType === Node.ELEMENT_NODE) {
-                        valueLabel.textContent = format(defaultPhysics[cfg.key]);
-                    }
-                });
-            });
-
-            actions.appendChild(resetBtn);
-            panel.appendChild(actions);
-
-            gameContainer.appendChild(panel);
-            return panel;
-        };
-
-        const physicsPanel = createPhysicsPanel();
+        const physicsPanel = createPhysicsPanel(physics, defaultPhysics, sliderConfigs, makeValueFormatter, highImpactKeys, highImpactHints, arenaMeshes);
 
         const keys = {};
         const keyDownHandler = (e) => keys[e.key.toLowerCase()] = true;
@@ -702,7 +376,7 @@ export function initOfflineGame(scene, gameObjects, tournament) {
                 ballVZ = -ballVZ * physics.sideWallRestitution;
                 ballSpinY *= physics.sideWallSpinRetention;
             }
-
+dd
             // The net blocks low shots near the middle of the table.
             const insideNetHeight = gameObjects.ball.position.y - 0.15 < physics.tableTopY + physics.netHeight;
             const insideNetDepth = Math.abs(gameObjects.ball.position.z) < physics.tableHalfWidth;
