+++ frontend/src/pong/game/physicsConfig.js
@@ -1,0 +1,67 @@
+const defaultPhysics = {
+    initialSpeedX: 0.10,
+    initialLateralSpeedZ: 0.013,
+    initialUpwardSpeedY: 0.09,
+    speedMultiplier: 0.999,
+    speedMax: 0.24,
+    gravity: 0.0042,
+    magnusCoefficient: 0.0008,
+    magnusExtraFactor: 20,
+    spinDecay: 0.992,
+    visualSpinFactor: 100,
+    tableTopY: 0,
+    tableHalfLength: 5.9,
+    tableHalfWidth: 2.95,
+    // Table is slightly higher at center and lower toward each end.
+    tableInclinePerUnitX: 0.01,
+    tableDownhillAcceleration: 0.0016,
+    tableBounceRestitution: 0.9,
+    tableBounceFriction: 0.985,
+    sideWallRestitution: 0.92,
+    sideWallSpinRetention: 0.9,
+    // Net height controls the center barrier visually and physically.
+    netHeight: 0.2,
+    netHalfThickness: 0.05,
+    netBounceDamping: 0.6,
+    netSpinRetention: 0.75,
+    floorY: -2,
+    goalX: 6.3,
+    // Collision dimensions are tuned to match the visible paddle face,
+    // plus a small margin for the ball radius.
+    paddleCollisionDepth: 0.2,
+    paddleCollisionHeight: 0.9,
+    paddleCollisionWidth: 0.45,
+    // Sideways paddle motion feeds these spin/trajectory values on contact.
+    spinTransferCoefficient: 1.5,
+    velocityTransfer: 0.11,
+    paddleBounceBoost: 1.02,
+    paddleBaseY: 0.9,
+    paddleBoundsZ: 2.5,
+    paddleRotationStep: 0.06,
+    paddleRotationLimit: 0.65,
+    // Tilt and rotation speed can add vertical lift to the outgoing ball.
+    paddleTiltToVerticalVelocity: 0.12,
+    paddleApproachBoost: 0.14,
+    paddleRetreatDampingStrength: 0.2,
+    paddleRetreatDampingMin: 0.65,
+    paddleRotationLiftFromSpeed: 0.02,
+    paddleStepZ: 0.10,
+    paddleSwingReach: 1.62,
+    paddleSwingRotationBoost: 0.86,
+    paddleSwingDuration: 0.24,
+    paddleLobDuration: 0.2,
+    paddleLobLiftBoost: 0.12,
+    paddleLobSlowFactor: 0.5,
+    paddleLobSideDamping: 0.65,
+    paddleLobRotationBoost: 0.22,
+    paddleWingDuration: 0.22,
+    paddleWingReach: 0.38,
+    paddleWingSideReach: 0.42,
+    paddleWingLiftBoost: 0.16,
+    paddleWingSideBoost: 0.08,
+    paddleWingSpinBoost: 0.85,
+    paddleWingForwardFactor: 0.95,
+    paddleWingRotationBoost: 0.28,
+};
+
+export default defaultPhysics;
